Extract Unity process focus helpers
