+++ cli/dispatcher/shared-inputs-stamp.json
@@ -1,4 +1,4 @@
 {
   "schemaVersion": 1,
-  "sharedInputsHash": "4ffd3b6d8156471cab11096b0f52f3696937bdba"
+  "sharedInputsHash": "3ffc7a8c10ff0b5e13116a7f2576670389527199"
 }
