+++ cli/project-runner/shared-inputs-stamp.json
@@ -1,4 +1,4 @@
 {
   "schemaVersion": 1,
-  "sharedInputsHash": "552a3bbd8757b2357b7fca535453fdf2e416577c"
+  "sharedInputsHash": "43724cc079cfddb1a27f326ef31bfd7be980787e"
 }
